libcommute: Reworked versions of `Hamiltonian` and `HamiltonianPart`
`HamiltonianPart` is now templated on the matrix storage type (real or
complex elements), while `Hamiltonian` employs type erasure and stores
parts as a vector of `std::shared_ptr<void>`.

The Hamiltonian expression is not stored in `Hamiltonian`. Instead, it
is only passed to `prepare()` to construct a matrix representation.

Since the matrix representation is now generated by means of libcommute,
a new header `LibcommuteEigen.h` has been added to teach libcommute
how to use Eigen vectors and blocks as StateVector's.